requirements.txt
amqp==2.5.1
apipkg==1.5
asgiref==3.2.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
attrs-mate==0.0.5
autobahn==19.9.2
Automat==0.7.0
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
coverage==4.5.4
cryptography==2.7
Cython==0.29.13
daphne==2.3.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-debug-toolbar==2.0
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
execnet==1.7.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.20
incremental==17.5.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
kazoo==2.6.1
kombu==4.6.4
Markdown==3.1.1
more-itertools==7.2.0
mysqlclient==1.4.4
oauthlib==3.1.0
packaging==19.1
parso==0.5.1
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-forked==1.0.2
pytest-mock==1.10.4
pytest-xdist==1.29.0
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
zipp==0.6.0
zope.interface==4.6.0
